+++ styleframe/styler.py
@@ -1,4 +1,3 @@
-# coding:utf-8
 from . import utils
 from colour import Color
 from openpyxl.formatting.rule import ColorScaleRule
@@ -14,11 +13,11 @@ class Styler(object):
 
     cache = {}
 
-    def __init__(self, bg_color=None, bold=False, font=utils.fonts.arial, font_size=12, font_color=None,
+    def __init__(self, bg_color=None, bold=False, font=utils.fonts.arial, font_size=12.0, font_color=None,
                  number_format=utils.number_formats.general, protection=False, underline=None,
                  border_type=utils.borders.thin, horizontal_alignment=utils.horizontal_alignments.center,
                  vertical_alignment=utils.vertical_alignments.center, wrap_text=True, shrink_to_fit=True,
-                 fill_pattern_type=utils.fill_pattern_types.solid, indent=0, comment_author=None, comment_text=None,
+                 fill_pattern_type=utils.fill_pattern_types.solid, indent=0.0, comment_author=None, comment_text=None,
                  text_rotation=0):
 
         def get_color_from_string(color_str, default_color=None):
@@ -46,6 +45,12 @@ class Styler(object):
         self.comment_author = comment_author
         self.comment_text = comment_text
         self.text_rotation = text_rotation
+
+        if self.border_type == utils.borders.default_grid:
+            if self.bg_color is not None:
+                raise ValueError('bg_color and border_type={} can not be used together'.format(utils.borders.default_grid))
+            self.border_type = None
+            self.fill_pattern_type = None
 
     def __eq__(self, other):
         if not isinstance(other, self.__class__):